Remove unsafe RolePolicySpec.
This isn't as bad as it looks, whoever did it originally
at least hid everything internally (i.e. not a public
facing ssh). But it is still not ok. The box was shut
down when this was discovered but this is following up
and removing this test all together (it is well and
truly convered by the fact that all the other s3 tests
are run based on the same s3 policy anyway).